refactor: domain 내부 디렉토리 분리
