+++ src/main/java/com/kuit/conet/dao/TeamDao.java
@@ -1,7 +1,7 @@
 package com.kuit.conet.dao;
 
-import com.kuit.conet.domain.Team;
-import com.kuit.conet.domain.TeamMember;
+import com.kuit.conet.domain.team.Team;
+import com.kuit.conet.domain.team.TeamMember;
 import com.kuit.conet.dto.response.StorageImgResponse;
 import com.kuit.conet.dto.response.team.GetTeamResponse;
 import lombok.extern.slf4j.Slf4j;
