+++ src/main/java/com/kuit/conet/domain/plan/WaitingPlan.java
@@ -1,4 +1,4 @@
-package com.kuit.conet.domain;
+package com.kuit.conet.domain.plan;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
