+++ src/main/java/com/kuit/conet/domain/user/User.java
@@ -1,5 +1,6 @@
-package com.kuit.conet.domain;
+package com.kuit.conet.domain.user;
 
+import com.kuit.conet.domain.auth.Platform;
 import lombok.Getter;
 import lombok.Setter;
 import lombok.NoArgsConstructor;
