+++ src/main/java/com/kuit/conet/service/AuthService.java
@@ -7,8 +7,8 @@ import com.kuit.conet.auth.kakao.KakaoUserProvider;
 import com.kuit.conet.common.exception.InvalidTokenException;
 import com.kuit.conet.common.exception.UserException;
 import com.kuit.conet.dao.UserDao;
-import com.kuit.conet.domain.Platform;
-import com.kuit.conet.domain.User;
+import com.kuit.conet.domain.auth.Platform;
+import com.kuit.conet.domain.user.User;
 import com.kuit.conet.dto.request.auth.LoginRequest;
 import com.kuit.conet.dto.request.auth.PutOptionTermAndNameRequest;
 import com.kuit.conet.dto.response.auth.AgreeTermAndPutNameResponse;
