+++ src/main/java/com/kuit/conet/service/TeamService.java
@@ -3,9 +3,9 @@ package com.kuit.conet.service;
 import com.kuit.conet.common.exception.TeamException;
 import com.kuit.conet.dao.TeamDao;
 import com.kuit.conet.dao.UserDao;
-import com.kuit.conet.domain.StorageDomain;
-import com.kuit.conet.domain.Team;
-import com.kuit.conet.domain.TeamMember;
+import com.kuit.conet.domain.storage.StorageDomain;
+import com.kuit.conet.domain.team.Team;
+import com.kuit.conet.domain.team.TeamMember;
 import com.kuit.conet.dto.request.team.CreateTeamRequest;
 import com.kuit.conet.dto.request.team.ParticipateTeamRequest;
 import com.kuit.conet.dto.request.team.TeamIdRequest;
